+++ webapp/bower.json
@@ -8,6 +8,7 @@
     "leaflet": "^0.7.7",
     "ui-leaflet": "^1.0.0",
     "moment": "^2.13.0",
-    "angular-bootstrap": "^1.3.2"
+    "angular-bootstrap": "^1.3.2",
+    "angular-nvd3": "^1.0.8"
   }
 }
