+++ webapp/www/templates/metrics.html
@@ -1,0 +1,66 @@
+<ion-view view-title="Metrics">
+  <ion-content class="has-header padding">
+    <div class="row">
+      <div class="col col-25">
+      <button class="button button-stable button-block row"
+                ng-click="changeMetric()">{{selectCtrl.metricString}}</button>
+      </div>
+      <div class="col col-66">
+        <div class="row">
+          <div class="col">
+          <button class="button button-small" ng-disabled="true" style="text-align:right">From:</button>
+          </div><div class="col">
+          <input type="text" placeholder="year" ng-model="selectCtrl.fromDate.year">
+          </div><div class="col">
+          <input type="text" placeholder="month" ng-model="selectCtrl.fromDate.month">
+          </div><div class="col">
+          <!--
+          <input type="text" placeholder="day" ng-model="selectCtrl.fromDate.day">
+          -->
+          </div><div class="col">
+          <button class="button button-outline button-dark button-small"
+                  ng-click="changeFromWeekday()">{{selectCtrl.fromDateWeekdayString}}</button>
+          </div><div class="col">
+          <!--
+          <input type="text" placeholder="hour" ng-model="selectCtrl.fromDate.hour">
+          -->
+          </div>
+        </div>
+        <div class="row">
+          <div class="col">
+          <button class="button button-small" ng-disabled="true" class="col" style="text-align:right">&nbsp;&nbsp;&nbsp;To:</button>
+          </div><div class="col">
+          <input type="text" placeholder="year" ng-model="selectCtrl.toDate.year">
+          </div><div class="col">
+          <input type="text" placeholder="month" ng-model="selectCtrl.toDate.month">
+          </div><div class="col">
+          <!--
+          <input type="text" placeholder="day" ng-model="selectCtrl.toDate.day">
+          -->
+          </div><div class="col">
+          <button class="button button-outline button-dark button-small"
+                  ng-click="changeToWeekday()">{{selectCtrl.toDateWeekdayString}}</button>
+          </div><div class="col">
+          <!--
+          <input type="text" placeholder="hour" ng-model="selectCtrl.toDate.hour">
+          -->
+          </div> <!-- col-->
+        </div> <!-- row -->
+      </div> <!-- col -->
+      <div class="col col-10">
+        <button class="button button-assertive"
+                ng-click="getMetrics()">Get!</button>
+      </div>
+
+      <!--
+      <em-local-date date="selectCtrl.fromDate"></em-local-date>
+      <em-local-date date="selectCtrl.toDate"></em-local-date>
+      -->
+      <!--
+        <button style="" class="button button-stable button-block  icon ion-checkmark">Valid</button>
+      -->
+    </div> <!-- row -->
+    <!-- <nvd3 options="options" data="data" config="{refreshDataOnly: false}"></nvd3> -->
+    <nvd3 options="options" data="data"></nvd3>
+  </ion-content>
+</ion-view>
